FastMTU works. There's still some work to be done to integrate it into the config etc. but the basic algorithm works. (more detail in the description)
FastMTU starts with a fixed range, e.g. 0-2000Bytes. Then splits that
range in an amount of packets, e.g. 20. Those packets get sent out and
a timeout timer starts ticking. Once the timeout reaches zero the
gathered reponse packets are counted. The biggest successful
response/reply pair is taken as new baseline, aka rangeStart, the last
iterationStep e.g. 100 is added on top of that and then it’s set as
rangeEnd. The algorithm will run over that range again, until at some
point the iterationStep reach 1. Then we know the exact MTU. The
advantage of this is also that we know exactly how long  it will take
as long as the MTU is within certain bounds. E.g. to find the exact MTU
between 1-2000bytes takes 3 Iterations. 3 Iterations, contain 3
timeouts, so it takes 30sec if a timeout is set to 10s.

Of course this can be sped up if we set the timeout lower or we
increase the amount of packets sent at one time.